Trekk ut Bostedsadresse til egen komponent
